Style initial cards
