+++ src/Components/CardContainer/CardContainer.scss
@@ -1,7 +1,10 @@
-.generate-colors-btn {
-  background: purple;
+.cards-container {
+  display: flex;
+  flex-wrap: wrap;
 }
 
-.save-palette-btn {
-  background: orange;
+.buttons-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
 }
