+++ src/Components/ColorBox/ColorBox.scss
@@ -1,0 +1,12 @@
+.color-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.card {
+  margin: 5px;
+  p {
+    text-align: center;
+  }
+}
