+++ src/index.scss
@@ -1,2 +1,3 @@
 @import './Containers/App/App';
-@import './Components/CardContainer/CardContainer.scss'
+@import './Components/CardContainer/CardContainer';
+@import './Components/ColorBox/ColorBox';
